99% done, need to implement the correct templates and then we're good to go.
